requirements.txt
pyshorteners==1.0.1
python-dotenv==1.0.0
aiogram==2.25.1